requirements.txt
requests
pandas
numpy
yfinance
tqdm
lxml
line-bot-sdk
flask
twstock
html5lib